Dev: define api with 5 mins cache
