+++ .gitignore
@@ -4,4 +4,3 @@ yarn-error.log
 .DS_Store
 tsconfig.tsbuildinfo
 __auto-generated-denylist.json
-__auto-generated-status.json
