+++ vercel.json
@@ -12,11 +12,5 @@
         "Referrer-Policy": "origin-when-cross-origin"
       }
     }
-  ],
-  "crons": [
-    {
-      "path": "/api/cron/generate-status",
-      "schedule": "*/5 * * * *"
-    }
   ]
 }
